Setup Github Action to run SourceCred and Deploy site to Github Pages
This Action will save the SourceCred cache and invalidate it whenever the config files change, or the
SC version is updated. It runs the yarn load, graph, score, and site commands. After that, it packages the
files into the site folder and deploys it to github pages on a separate gh-pages branch.

Ideally we can just gitignore the site folder if we update the site command to package the output and sourcecred.json
files into the site folder. That way we can just get rid of all the symlinks as well.

Test Plan: Ensure that the action runs successfully and that the frontend is available on GH Pages