Add command to clean cache
Sometimes the cached data gets in a bad state which causes the sourcecred commands to fail.
This adds a command to package.json to make it easy for the user to wipe the cache, site and output folders
rimraf is used for cross-platform compatibility (i.e. Windows)

Test Plan: Run yarn clean and make sure the cache, output and site directories are deleted