Update Github Action for new site command + CNAME file copy
Site command now automatically copies the correct folders/files on its own, so we no longer
need to do it in the github action. Also added a step to copy a CNAME file from master to
gh-pages, if it exists.

Test Plan: Ensure the GH action runs successfully and deploys the GH pages site